create a class for StringFromDateProtocal
